+++ src/onegov.feriennet/onegov/feriennet/collections/activity.py
@@ -11,7 +11,8 @@ class VacationActivityCollection(ActivityCollection):
                  durations=None,
                  age_ranges=None,
                  owners=None,
-                 period_ids=None):
+                 period_ids=None,
+                 dateranges=None):
         super().__init__(
             session=session,
             type='vacation',
@@ -21,7 +22,8 @@ class VacationActivityCollection(ActivityCollection):
             durations=durations,
             age_ranges=age_ranges,
             owners=owners,
-            period_ids=period_ids
+            period_ids=period_ids,
+            dateranges=dateranges,
         )
         self.identity = identity
 
@@ -45,5 +47,6 @@ class VacationActivityCollection(ActivityCollection):
             durations=self.durations,
             age_ranges=self.age_ranges,
             owners=self.owners,
-            period_ids=self.period_ids
+            period_ids=self.period_ids,
+            dateranges=self.dateranges
         )